requirements.txt
absl-py==1.4.0
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
Deprecated==1.2.14
docstring-parser==0.15
fire==0.5.0
google-api-core==2.14.0
google-api-python-client==1.12.11
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-pipeline-components==0.1.1
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kfp==1.8.22
kfp-pipeline-spec==0.1.16
kfp-server-api==1.8.5
kubernetes==25.3.0
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.4.0
shapely==2.0.2
six==1.16.0
strip-hints==0.1.10
tabulate==0.9.0
termcolor==2.4.0
typer==0.9.0
typing-extensions==4.8.0
uritemplate==3.0.1
urllib3==1.26.18
websocket-client==1.7.0
wrapt==1.16.0
xgboost==2.0.3
fsspec==2024.2.0
gcsfs==2024.2.0